+++ nacossync-worker/src/main/java/com/alibaba/nacossync/constant/SkyWalkerConstants.java
@@ -38,4 +38,6 @@ public class SkyWalkerConstants {
     public static final String SYNC_INSTANCE_TAG="sync.instance.tag";
     public static final Integer MAX_THREAD_NUM = 200;
 
+    public static final String ALL = "ALL";
+
 }
