+++ nacossync-worker/src/main/java/com/alibaba/nacossync/pojo/request/CatalogServiceResult.java
@@ -1,0 +1,36 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.alibaba.nacossync.pojo.request;
+
+import com.alibaba.nacossync.pojo.view.ServiceView;
+import lombok.Data;
+
+import java.util.List;
+
+/**
+ * @author NacosSync
+ * @since 2024-01-04 15:54:20
+ */
+@Data
+public class CatalogServiceResult {
+    /**
+     * count，not equal serviceList.size .
+     */
+    private int count;
+
+    private List<ServiceView> serviceList;
+}
