+++ nacossync-worker/src/main/java/com/alibaba/nacossync/pojo/view/ServiceView.java
@@ -1,0 +1,38 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.alibaba.nacossync.pojo.view;
+
+import lombok.Data;
+
+/**
+ * @author NacosSync
+ * @since 2024-01-04 15:54:36
+ */
+@Data
+public class ServiceView {
+    private String name;
+
+    private String groupName;
+
+    private int clusterCount;
+
+    private int ipCount;
+
+    private int healthyInstanceCount;
+
+    private String triggerFlag;
+}
